requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asttokens==2.4.1
astunparse==1.6.3
blosc2==2.7.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
ipdb==0.13.13
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
ndindex==1.9.2
networkx==3.4.2
numexpr==2.10.2
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.29.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-logger==0.3.1
setuptools==75.6.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tables==3.10.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
torch==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
visualize==0.5.1
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
